+++ cansocket/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Fix frame send data length
+- Use libc constants for error frame checking
 
 ## v0.1.0